+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `categoryFiltersMode` prop to Filter Navigator Flexible.
 
 ## [3.86.4] - 2020-12-02
 ### Fixed